pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "glm-agent"
version = "0.1.0"
description = "AI Agent built with Zhipu GLM-4.7"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Kanno Kouki", email = "kouki0180@outlook.jp" }
]

dependencies = [
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
glm-agent = "glm_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/glm_agent"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
